net: http: Remove the old legacy API
There are no internal users for old HTTP API so removing it.
