+++ subsys/net/ip/net_shell.c
@@ -1398,37 +1398,6 @@ static char *http_str_output(char *output, int outlen, const char *str, int len)
 	return output;
 }
 
-#if !defined(CONFIG_HTTP_APP)
-static void http_server_cb(struct http_server_ctx *entry,
-			   void *user_data)
-{
-	int *count = user_data;
-	static char output[MAX_HTTP_OUTPUT_LEN];
-
-	/* +7 for []:port */
-	char addr_local[ADDR_LEN + 7];
-	char addr_remote[ADDR_LEN + 7] = "";
-
-	get_addresses(entry->req.net_ctx, addr_local, sizeof(addr_local),
-		      addr_remote, sizeof(addr_remote));
-
-	if (*count == 0) {
-		printk("        HTTP ctx    Local           \t"
-		       "Remote          \tURL\n");
-	}
-
-	(*count)++;
-
-	printk("[%2d] %c%c %p  %16s\t%16s\t%s\n",
-	       *count, entry->enabled ? 'E' : 'D',
-	       entry->is_https ? 'S' : ' ',
-	       entry, addr_local, addr_remote,
-	       http_str_output(output, sizeof(output) - 1,
-			       entry->req.url, entry->req.url_len));
-}
-#endif
-
-#if defined(CONFIG_HTTP_APP)
 static void http_server_cb(struct http_ctx *entry, void *user_data)
 {
 	int *count = user_data;
@@ -1465,7 +1434,6 @@ static void http_server_cb(struct http_ctx *entry, void *user_data)
 				       entry->http.url, entry->http.url_len));
 	}
 }
-#endif /* CONFIG_HTTP_APP */
 #endif /* CONFIG_NET_DEBUG_HTTP_CONN && CONFIG_HTTP_SERVER */
 
 int net_shell_cmd_http(int argc, char *argv[])
