+++ subsys/net/lib/http/CMakeLists.txt
@@ -7,15 +7,10 @@ endif()
 zephyr_library_sources_if_kconfig(http_parser.c)
 zephyr_library_sources_if_kconfig(http_parser_url.c)
 
-if(CONFIG_HTTP_APP)
-  zephyr_library_sources(http_app.c)
+zephyr_library_sources(http.c)
 
-  zephyr_library_sources_ifdef(CONFIG_HTTP_SERVER http_app_server.c)
-  zephyr_library_sources_ifdef(CONFIG_HTTP_CLIENT http_app_client.c)
-else()
-  zephyr_library_sources_ifdef(CONFIG_HTTP_SERVER http_server.c)
-  zephyr_library_sources_ifdef(CONFIG_HTTP_CLIENT http_client.c)
-endif()
+zephyr_library_sources_ifdef(CONFIG_HTTP_SERVER http_server.c)
+zephyr_library_sources_ifdef(CONFIG_HTTP_CLIENT http_client.c)
 
 zephyr_link_interface_ifdef(CONFIG_MBEDTLS mbedTLS)
 zephyr_library_link_libraries_ifdef(CONFIG_MBEDTLS mbedTLS)
